+++ src/callbacksCustom/preItemPickup.ts
@@ -1,9 +1,8 @@
 import { PickingUpItem } from "isaacscript-common";
 import * as streakText from "../features/mandatory/streakText";
+import * as automaticItemInsertion from "../features/optional/quality/automaticItemInsertion/automaticItemInsertion";
 
-export function main(
-  _player: EntityPlayer,
-  pickingUpItem: PickingUpItem,
-): void {
+export function main(player: EntityPlayer, pickingUpItem: PickingUpItem): void {
   streakText.preItemPickup(pickingUpItem);
+  automaticItemInsertion.preItemPickup(player, pickingUpItem);
 }